+++ voxelbotutils/cogs/command_event.py
@@ -17,7 +17,12 @@ class CommandEvent(vbu.Cog):
         content = ctx.message.content.replace('\n', '\\n')[:self.CONTENT_LIMIT]
         if len(ctx.message.content) > self.CONTENT_LIMIT:
             content += '...'
-        invoke_text = "Interaction invoked" if ctx.is_interaction else "Command invoked"
+        invoke_text = "Command invoked"
+        if ctx.is_interaction:
+            if getattr(ctx, "given_values", None) is not None:
+                invoke_text = "Context invoked"
+            else:
+                invoke_text = "Interaction invoked"
         if ctx.guild is None:
             return logger.info(f"{invoke_text} ({ctx.invoked_with}) ~ (G0/C{ctx.channel.id}/U{ctx.author.id}) :: {content}")
         logger.info(f"{invoke_text} ({ctx.invoked_with}) ~ (G{ctx.guild.id}/C{ctx.channel.id}/U{ctx.author.id}) :: {content}")